requirements.txt
fastapi
uvicorn
httpx
replicate
python-dotenv
python-multipart
